back/requirements.txt
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.7
contourpy==1.1.0
cryptography==41.0.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-seed==0.2.2
djangorestframework==3.14.0
django-environ==0.11.2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
Faker==19.11.0
fastjsonschema==2.17.1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.41.1
fqdn==1.5.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
kiwisolver==1.4.4
locust==2.17.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
msgpack==1.0.7
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==7.0.0
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
rpds-py==0.9.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
widgetsnbextension==4.0.8
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
